working through errors

+++ src/audio/playback.rs
@@ -357,11 +357,11 @@ impl AudioSource for PlaybackDevice {
 
             ControlAction::TransportJumpBeat(beat_idx) => {
                 (self.current_clip, self.active, self.current_sample) =
-                    self.calculate_time_at_beat(ctx, beat_idx);
+                    self.calculate_time_at_beat(ctx, beat_idx as u16);
             }
             ControlAction::TransportSeekBeat(beat_idx) => {
                 (self.current_clip, self.active, self.current_sample) =
-                    self.calculate_time_at_beat(ctx, beat_idx);
+                    self.calculate_time_at_beat(ctx, beat_idx as u16);
                 // TODO: Support multiple and mixed sample rates
                 self.current_sample -= (ctx.transport.us_to_next_beat as i32) * 48 / 1000
             }
